- Deleted unused images from res drawable - Implemented navigation for all bottom-bar-navigation-icons except for "Board" - Deleted unused imports in ShiftsScreen - Moved ShiftsScreenContent composable to into separate file
