+++ app/src/main/java/com/example/stablestable/navigation/Navigation.kt
@@ -42,10 +42,10 @@ fun SetupNavGraph(
         ) {
             HomeScreen(
                 goToRiders = { navController.navigate(Screen.StableUsers.route) },
-                goToHome = { navController.navigate(Screen.HomeScreen.route) }
-                goToShifts = { navController.navigate(Screen.StableUsers.route) },
+                goToHome = { navController.navigate(Screen.HomeScreen.route) },
+                goToShifts = { navController.navigate(Screen.ShiftsScreen.route) },
                 //goToProfile = { navController.navigate(Screen.UserProfileScreen.route.replace("{userId}", authViewModel.userId ?: "")) },
-                //goToHorses = { navController.navigate(Screen.StableHorses.route) },
+                goToHorses = { navController.navigate(Screen.StableHorses.route) }
                 //onLogout = { authViewModel.setUserLoggedIn(false) }
             )
         }
@@ -59,7 +59,9 @@ fun SetupNavGraph(
                         userId -> navController.navigate(Screen.UserProfileScreen.route.replace("{userId}", userId))
                 },
                 goToRiders = { navController.navigate(Screen.StableUsers.route) },
-                goToHome = { navController.navigate(Screen.HomeScreen.route) }
+                goToHome = { navController.navigate(Screen.HomeScreen.route) },
+                goToShifts = { navController.navigate(Screen.ShiftsScreen.route) },
+                goToHorses = { navController.navigate(Screen.StableHorses.route) }
             )
         }
 
@@ -72,7 +74,10 @@ fun SetupNavGraph(
                         horseId -> navController.navigate(Screen.HorseProfileScreen.route.replace("{horseId}", horseId))
                 },
                 goToRiders = { navController.navigate(Screen.StableUsers.route) },
-                goToHome = { navController.navigate(Screen.HomeScreen.route) }
+                goToHome = { navController.navigate(Screen.HomeScreen.route) },
+                goToShifts = { navController.navigate(Screen.ShiftsScreen.route) },
+                goToHorses = { navController.navigate(Screen.StableHorses.route) }
+
             )
         }
 
@@ -87,7 +92,10 @@ fun SetupNavGraph(
                         horseId -> navController.navigate(Screen.HorseProfileScreen.route.replace("{horseId}", horseId))
                 },
                 goToRiders = { navController.navigate(Screen.StableUsers.route) },
-                goToHome = { navController.navigate(Screen.HomeScreen.route) }
+                goToHome = { navController.navigate(Screen.HomeScreen.route) },
+                goToShifts = { navController.navigate(Screen.ShiftsScreen.route) },
+                goToHorses = { navController.navigate(Screen.StableHorses.route) }
+
             )
         }
 
@@ -98,8 +106,11 @@ fun SetupNavGraph(
         ) { backStackEntry ->
             HorseProfileScreen(
                 horseId = backStackEntry.arguments?.getString("horseId") ?: "",
-            goToRiders = { navController.navigate(Screen.StableUsers.route) },
-            goToHome = { navController.navigate(Screen.HomeScreen.route) }
+                goToRiders = { navController.navigate(Screen.StableUsers.route) },
+                goToHome = { navController.navigate(Screen.HomeScreen.route) },
+                goToShifts = { navController.navigate(Screen.ShiftsScreen.route) },
+                goToHorses = { navController.navigate(Screen.StableHorses.route) }
+
             )
         }
         //Route to shifts
@@ -107,7 +118,10 @@ fun SetupNavGraph(
             route = Screen.ShiftsScreen.route
         ){
             ShiftsScreen(
-                goToHomeScreen = { navController.navigate(Screen.HomeScreen.route)}
+                goToHome = { navController.navigate(Screen.HomeScreen.route)},
+                goToRiders = { navController.navigate(Screen.StableUsers.route) },
+                goToShifts = { navController.navigate(Screen.ShiftsScreen.route) },
+                goToHorses = { navController.navigate(Screen.StableHorses.route) }
             )
         }
     }
