+++ app/src/main/java/com/example/stablestable/navigation/Screen.kt
@@ -7,8 +7,6 @@ sealed class Screen(val route: String) {
     object UserProfileScreen: Screen(route = "userProfile/{userId}")
     object StableHorses: Screen(route = "stableHorses_screen")
     object HorseProfileScreen: Screen(route = "horseProfile/{horseId}")
-    object StableScreen: Screen(route = "stable_screen")
-    object CalendarScreen: Screen(route = "calendar_screen")
     object ShiftsScreen: Screen(route = "shifts_screen")
 
 }
