+++ app/src/main/java/com/example/stablestable/ui/stable/horses/StableHorsesScreen.kt
@@ -2,19 +2,22 @@ package com.example.stablestable.ui.stable.horses
 
 import androidx.compose.runtime.Composable
 import com.example.stablestable.components.CreateScaffold
-import com.example.stablestable.ui.stable.riders.StableUsersScreenContent
 
 @Composable
 fun StableHorsesScreen(
     goToRiders: () -> Unit,
     goToHome: () -> Unit,
+    goToShifts: () -> Unit,
+    goToHorses: () -> Unit,
     onHorseClick: (String) -> Unit
 ){
     CreateScaffold(
         content = {paddingValues ->
-            StableUsersScreenContent(paddingValues, onHorseClick)
+            StableHorsesScreenContent(paddingValues, onHorseClick)
         },
         goToHome = goToHome,
-        goToRiders = goToRiders
+        goToRiders = goToRiders,
+        goToShifts = goToShifts,
+        goToHorses = goToHorses
     )
 }
